added plotting with plotting funcitons and classes
